GSTRB-Full Stack Application/amazon_deployment/requirements.txt
pandas==0.25.1
numpy==1.17.2
Flask==1.0.2
h5py==2.9.0
Werkzeug==0.15.5
Keras==2.2.5
flask_restplus==0.13.0
Pillow==6.2.0
matplotlib==3.1.1
scikit-image==0.14.2
scikit_learn==0.21.3
tensorflow==2.0.0
